+++ assets/simple hex flat top.tsx
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.10.2" name="simple hex flat top" tilewidth="60" tileheight="52" tilecount="6" columns="0">
+ <grid orientation="orthogonal" width="1" height="1"/>
+ <tile id="1">
+  <image width="60" height="52" source="tiles/simple hex flat top/boat.png"/>
+ </tile>
+ <tile id="2">
+  <image width="60" height="52" source="tiles/simple hex flat top/cave.png"/>
+ </tile>
+ <tile id="4">
+  <image width="60" height="52" source="tiles/simple hex flat top/moutain.png"/>
+ </tile>
+ <tile id="5">
+  <image width="60" height="52" source="tiles/simple hex flat top/ocean.png"/>
+ </tile>
+ <tile id="6">
+  <image width="60" height="52" source="tiles/simple hex flat top/plain.png"/>
+ </tile>
+ <tile id="7">
+  <image width="60" height="52" source="tiles/simple hex flat top/forest.png"/>
+ </tile>
+</tileset>